requirements_flask.txt
# Flask and extensions
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug>=3.1.0

# Database
SQLAlchemy==2.0.46
PyMySQL==1.1.0

# Authentication & Security
python-dotenv>=1.0.0
PyJWT>=2.10.1

# LLM APIs for Code Generation
openai<2.0.0,>=1.68.2
google-generativeai>=0.4.0,<0.8
anthropic>=0.41.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Optional: for production
gunicorn==21.2.0
python-multipart>=0.0.18
